requirements.txt
torch
torchvision
transformers
pillow
weaviate-client>=3.26.0
fastapi
uvicorn
python-multipart
